refactor: 도메인들을 core 모듈로 이동
