+++ kitchenpos-core/build.gradle
@@ -1,0 +1,5 @@
+dependencies {
+    testImplementation 'org.junit.jupiter:junit-jupiter'
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+    testImplementation 'org.assertj:assertj-core:3.21.0'
+}
