+++ kitchenpos-core/src/test/java/menu/domain/MenuTest.java
@@ -1,6 +1,9 @@
-package kitchenpos.menu.domain;
+package menu.domain;
 
+import kitchenpos.menu.domain.Menu;
 import kitchenpos.product.domain.Product;
+import org.assertj.core.api.Assertions;
+import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
@@ -8,9 +11,6 @@ import org.junit.jupiter.api.Test;
 import java.math.BigDecimal;
 import java.util.Collections;
 import java.util.List;
-
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.assertj.core.api.SoftAssertions.assertSoftly;
 
 class MenuTest {
 
@@ -26,13 +26,13 @@ class MenuTest {
             final BigDecimal price = new BigDecimal(1000);
             final long menuGroupId = 1L;
             final Product product = new Product("product", new BigDecimal(3000));
-            final List<MenuProduct> menuProducts = List.of(new MenuProduct(product, 4));
+            final List<kitchenpos.menu.domain.MenuProduct> menuProducts = List.of(new kitchenpos.menu.domain.MenuProduct(product, 4));
 
             //when
             final Menu menu = new Menu(menuName, price, menuGroupId, menuProducts);
 
             //then
-            assertSoftly(softly -> {
+            SoftAssertions.assertSoftly(softly -> {
                 softly.assertThat(menu.getId()).isNull();
                 softly.assertThat(menu.getName()).isEqualTo(menuName);
                 softly.assertThat(menu.getPrice()).isEqualByComparingTo(price);
@@ -48,10 +48,10 @@ class MenuTest {
             final String menuName = "menu";
             final BigDecimal price = new BigDecimal(-1);
             final long menuGroupId = 1L;
-            final List<MenuProduct> menuProducts = Collections.emptyList();
+            final List<kitchenpos.menu.domain.MenuProduct> menuProducts = Collections.emptyList();
 
             // when & then
-            assertThatThrownBy(() -> new Menu(menuName, price, menuGroupId, menuProducts))
+            Assertions.assertThatThrownBy(() -> new Menu(menuName, price, menuGroupId, menuProducts))
                     .isInstanceOf(IllegalArgumentException.class);
         }
 
@@ -62,10 +62,10 @@ class MenuTest {
             final String menuName = "menu";
             final BigDecimal price = null;
             final long menuGroupId = 1L;
-            final List<MenuProduct> menuProducts = Collections.emptyList();
+            final List<kitchenpos.menu.domain.MenuProduct> menuProducts = Collections.emptyList();
 
             // when & then
-            assertThatThrownBy(() -> new Menu(menuName, price, menuGroupId, menuProducts))
+            Assertions.assertThatThrownBy(() -> new Menu(menuName, price, menuGroupId, menuProducts))
                     .isInstanceOf(IllegalArgumentException.class);
         }
     }
