+++ kitchenpos-core/src/test/java/order/domain/OrderTest.java
@@ -1,14 +1,15 @@
-package kitchenpos.order.domain;
+package order.domain;
 
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderStatus;
+import org.assertj.core.api.Assertions;
+import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
 import java.time.LocalDateTime;
 import java.util.Collections;
-
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.assertj.core.api.SoftAssertions.assertSoftly;
 
 class OrderTest {
 
@@ -23,7 +24,7 @@ class OrderTest {
             final Order order = new Order(null, OrderStatus.MEAL.name(), LocalDateTime.now(), Collections.emptyList());
 
             //then
-            assertSoftly(softly -> {
+            SoftAssertions.assertSoftly(softly -> {
                 softly.assertThat(order.getId()).isNull();
                 softly.assertThat(order.getOrderStatus()).isEqualTo(OrderStatus.MEAL.name());
             });
@@ -45,7 +46,7 @@ class OrderTest {
             order.changeOrderStatus(afterStatus);
 
             //then
-            assertSoftly(softly -> {
+            SoftAssertions.assertSoftly(softly -> {
                 softly.assertThat(order.getOrderStatus()).isEqualTo(afterStatus);
             });
         }
@@ -57,7 +58,7 @@ class OrderTest {
             final Order order = new Order(null, OrderStatus.COMPLETION.name(), LocalDateTime.now(), Collections.emptyList());
 
             // when & then
-            assertThatThrownBy(() -> order.changeOrderStatus(OrderStatus.MEAL.name()))
+            Assertions.assertThatThrownBy(() -> order.changeOrderStatus(OrderStatus.MEAL.name()))
                     .isInstanceOf(IllegalArgumentException.class);
         }
     }
