+++ settings.gradle
@@ -1,1 +1,3 @@
 rootProject.name = 'kitchenpos'
+include 'kitchenpos-core'
+
