--- src/main/java/kitchenpos/table/domain/OrderTableRepository.java
@@ -1,9 +1,0 @@
-package kitchenpos.table.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface OrderTableRepository extends JpaRepository<OrderTable, Long> {
-    List<OrderTable> findAllByIdIn(final List<Long> orderTableIds);
-}
